refactor: Moved our routes out from /embedding
